Remove unnecessary Line Endings
In alt_inline and alt_inline_attach cases, there are 3 $LE in total between the file attachments and the end of the boundary: at the end of the attachAll(), in the send code, and in the beginning of endBoundary(). After removing the extra $LE, I've sent test emails to Gmail and Yahoo. Everything worked fine.